requirements.txt
openai-whisper
torch
ffmpeg-python
streamlit